+++ dev/20190710/k2works/api/main.py
@@ -1,5 +1,5 @@
 import os
-from rental.app import app
+from video_rental.app import app
 
 if __name__ == '__main__':
     app.run(host=os.getenv('APP_ADDRESS', 'localhost'), port=8000)
