+++ dev/20190710/k2works/api/video_rental/domain/customer.py
@@ -1,4 +1,4 @@
-from .movie import Movie
+from movie import Movie
 
 
 class Customer:
